+++ SpatialGDK/Source/SpatialGDK/Private/Interop/Connection/SpatialWorkerConnection.cpp
@@ -159,25 +159,30 @@ void USpatialWorkerConnection::Advance()
 bool USpatialWorkerConnection::HasDisconnected() const
 {
 	check(Coordinator.IsValid());
-	return Coordinator->GetViewDelta().HasDisconnected();
+	return Coordinator->GetViewDelta().HasConnectionStatusChanged();
 }
 
 Worker_ConnectionStatusCode USpatialWorkerConnection::GetConnectionStatus() const
 {
 	check(Coordinator.IsValid());
-	return Coordinator->GetViewDelta().GetConnectionStatus();
+	return Coordinator->GetViewDelta().GetConnectionStatusChange();
 }
 
 FString USpatialWorkerConnection::GetDisconnectReason() const
 {
 	check(Coordinator.IsValid());
-	return Coordinator->GetViewDelta().GetDisconnectReason();
+	return Coordinator->GetViewDelta().GetConnectionStatusChangeMessage();
 }
 
 const SpatialGDK::EntityView& USpatialWorkerConnection::GetView() const
 {
 	check(Coordinator.IsValid());
 	return Coordinator->GetView();
+}
+
+SpatialGDK::ViewCoordinator& USpatialWorkerConnection::GetCoordinator() const
+{
+	return *Coordinator;
 }
 
 PhysicalWorkerName USpatialWorkerConnection::GetWorkerId() const