+++ SpatialGDK/Source/SpatialGDK/Public/Interop/SpatialReceiver.h
@@ -231,8 +231,6 @@ private:
 	UPROPERTY()
 	UGlobalStateManager* GlobalStateManager;
 
-	SpatialLoadBalanceEnforcer* LoadBalanceEnforcer;
-
 	FTimerManager* TimerManager;
 
 	SpatialGDK::SpatialRPCService* RPCService;
@@ -289,4 +287,5 @@ private:
 	bool HasEntityBeenRequestedForDelete(Worker_EntityId EntityId);
 	void HandleDeferredEntityDeletion(const DeferredRetire& Retire);
 	void HandleEntityDeletedAuthority(Worker_EntityId EntityId);
+	bool IsDynamicSubObject(AActor* Actor, uint32 SubObjectOffset);
 };