+++ SpatialGDK/Source/SpatialGDK/Public/Utils/SpatialStatics.h
@@ -9,6 +9,7 @@
 #include "UObject/TextProperty.h"
 
 #include "SpatialGDKSettings.h"
+#include "Utils/SpatialDebugger.h"
 
 #include "SpatialStatics.generated.h"
 
@@ -163,6 +164,9 @@ public:
 	 */
 	UFUNCTION(BlueprintCallable, BlueprintPure, Category = "SpatialOS", meta = (WorldContext = "WorldContextObject"))
 	static FName GetLayerName(const UObject* WorldContextObject);
+
+	UFUNCTION(BlueprintCallable, Category = "SpatialGDK|Spatial Debugger", meta = (WorldContext = "WorldContextObject"))
+	static void SpatialDebuggerSetOnConfigUIClosedCallback(const UObject* WorldContextObject, FOnConfigUIClosedDelegate Delegate);
 
 private:
 	static FName GetCurrentWorkerType(const UObject* WorldContext);